fix(tooltip): Fix tooltip y coordinate with arc.rangeText
- Adjust tooltip y coordinate position with combination of
arc.rangeText + title.text in gauge type
- Split tooltip position related test in separate file

Ref #3673
